+++ src/Avatar.h
@@ -142,7 +142,7 @@ public:
 	bool questlog_dismissed;
 	bool using_main1;
 	bool using_main2;
-	int prev_hp;
+	float prev_hp;
 	bool playing_lowhp;
 	bool teleport_camera_lock;
 	int feet_index;
