+++ src/Hazard.h
@@ -57,10 +57,10 @@ public:
 	bool relative_pos;
 	bool sfx_hit_played;
 
-	int dmg_min;
-	int dmg_max;
-	int crit_chance;
-	int accuracy;
+	float dmg_min;
+	float dmg_max;
+	float crit_chance;
+	float accuracy;
 	int source_type;
 	float base_speed;
 	int lifespan; // ticks down to zero
