+++ src/LootManager.cpp
@@ -743,7 +743,7 @@ void LootManager::checkLootComponent(EventComponent* ec, FPoint *pos, std::vecto
 	// an item id of 0 means we should drop currency instead
 	if (ec->id == 0 || ec->id == eset->misc.currency_id) {
 		new_loot.item = eset->misc.currency_id;
-		new_loot.quantity = new_loot.quantity * (100 + pc->stats.get(Stats::CURRENCY_FIND)) / 100;
+		new_loot.quantity = static_cast<int>(static_cast<float>(new_loot.quantity) * (100 + pc->stats.get(Stats::CURRENCY_FIND)) / 100);
 	}
 	else {
 		new_loot.item = ec->id;
